+++ Web/Project_Footprint/website/urls.py
@@ -1,5 +1,6 @@
 from django.urls import path
 from . import views
+from . import viewsets
 
 urlpatterns = [
     path('', views.index, name='index'),
@@ -15,10 +16,10 @@ urlpatterns = [
     path('place_sights_list/', views.place_sights, name='place_sights_list'),
     path('index/', views.index, name='index'),
     path('place_search/',views.place_search,name='place_search'),
-    path('history/update/',views.history_update,name='history-update'),
-    path('history/create/',views.history_create,name='history-create'),
-    path('history/<int:id>/delete/', views.history_delete,name='history-delete'),
-    path('user_info_update/', views.user_info_update, name='user_info_update'),
-    path('user_delete/', views.user_delete, name='user_delete'),
+    path('history/update/',viewsets.history_update,name='history-update'),
+    path('history/create/',viewsets.history_create,name='history-create'),
+    path('history/<int:id>/delete/', viewsets.history_delete,name='history-delete'),
+    path('user_info_update/', viewsets.user_info_update, name='user_info_update'),
+    path('user_delete/', viewsets.user_delete, name='user_delete'),
     path('myinfo/', views.myinfo, name='myinfo')
 ]
